consolidate syscall-related code; simplify initialization
