+++ internal/channelz/syscall_test.go
@@ -1,5 +1,4 @@
 //go:build linux
-// +build linux
 
 /*
  *
